added the final fixes for task 6

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int b;
 long int c;
 long long int d;
 float e;
-	printf("size of a char: %lu byte(s)", sizeof(a));
-	printf("\nsize of an int: %lu byte(s)", sizeof(b));
-	printf("\nsize of a long int: %lu byte(s)", sizeof(c));
-	printf("\nsize of a long long int: %lu byte(s)", sizeof(d));
-	printf("\nsize of a float: %lu byte(s)\n", sizeof(e));
+	printf("size of a char: %lu byte(s)", (unsigned long) sizeof(a));
+	printf("\nsize of an int: %lu byte(s)", (unsigned long) sizeof(b));
+	printf("\nsize of a long int: %lu byte(s)", (unsigned long)sizeof(c));
+	printf("\nsize of a long long int: %lu byte(s)", (unsigned long)sizeof(d));
+	printf("\nsize of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
 	return (0);
 }
